model/requirements.txt
imageio
jupyter
matplotlib
numpy
pillow
pytest
pyyaml
scipy